backend/requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
boto3==1.34.69
botocore==1.34.69
cattrs==23.2.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
contourpy==1.2.1
coverage==7.5.0
cryptography==42.0.5
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1
django-admin-honeypot==1.1.0
django-allauth==64.2.0
django-anymail==10.3
django-appconf==1.0.6
django-filter==23.5
django-hosts==6.0
django-modelcluster==6.3
django-permissionedforms==0.1
django-select2==8.2.0
django-storages==1.14.2
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
fido2==1.1.3
filetype==1.2.0
flake8==7.0.0
fonttools==4.53.0
geojson-rewind==1.1.0
geomet==1.1.0
git-filter-repo==2.45.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.8
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
kiwisolver==1.4.5
l18n==2021.3
laces==0.1.1
libusb1==3.1.0
llm==0.15
MarkupSafe==2.1.5
matplotlib==3.9.0
mccabe==0.7.0
ndeflib==0.3.3
nfcpy==1.0.4
numpy==1.26.4
oauthlib==3.2.2
openai==1.41.1
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pillow==10.4.0
pillow_heif==0.15.0
pip-chill==1.0.3
platformdirs==4.2.2
pluggy==1.5.0
promise==2.3
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pyDes==2.0.1
pyflakes==3.2.0
pygbif==0.6.4
PyJWT==2.8.0
pyparsing==3.1.2
pypng==0.20220715.0
pyserial==3.5
PySocks==1.7.1
pytest==8.2.0
pytest-base-url==2.1.0
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-selenium==4.1.0
pytest-variables==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ulid==2.7.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
queryish==0.2
requests==2.32.3
requests-cache==1.2.0
requests-oauthlib==2.0.0
s3transfer==0.10.1
selenium==4.20.0
setuptools==73.0.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.37
sqlparse==0.5.1
tabulate==0.9.0
telepath==0.3.1
tenacity==8.3.0
text-unidecode==1.3
tqdm==4.66.5
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
url-normalize==1.4.3
urllib3==2.2.2
wagtail==6.2.1
wagtail-ai==2.1.1
webencodings==0.5.1
whitenoise==6.6.0
Willow==1.8.0
wsproto==1.2.0
